ibcutils/utils_test.go
package ibcutils

import (
	"testing"

	"github.com/stretchr/testify/require"

	sdk "github.com/cosmos/cosmos-sdk/types"

	transfertypes "github.com/cosmos/ibc-go/v4/modules/apps/transfer/types"
	channeltypes "github.com/cosmos/ibc-go/v4/modules/core/04-channel/types"
	ibctesting "github.com/cosmos/ibc-go/v4/testing"
)

func init() {
	cfg := sdk.GetConfig()
	cfg.SetBech32PrefixForAccount("althea", "altheapub")
}

func TestGetTransferSenderRecipient(t *testing.T) {
	testCases := []struct {
		name         string
		packet       channeltypes.Packet
		expSender    string
		expRecipient string
		expError     bool
	}{
		{
			"empty packet",
			// nolint: exhaustruct
			channeltypes.Packet{},
			"", "",
			true,
		},
		{
			"invalid packet data",
			// nolint: exhaustruct
			channeltypes.Packet{
				Data: ibctesting.MockFailPacketData,
			},
			"", "",
			true,
		},
		{
			"empty FungibleTokenPacketData",
			// nolint: exhaustruct
			channeltypes.Packet{
				Data: transfertypes.ModuleCdc.MustMarshalJSON(
					// nolint: exhaustruct
					&transfertypes.FungibleTokenPacketData{},
				),
			},
			"", "",
			true,
		},
		{
			"invalid sender",
			// nolint: exhaustruct
			channeltypes.Packet{
				Data: transfertypes.ModuleCdc.MustMarshalJSON(
					// nolint: exhaustruct
					&transfertypes.FungibleTokenPacketData{
						Sender:   "cosmos1",
						Receiver: "althea1x2w87cvt5mqjncav4lxy8yfreynn273x9h93zp",
						Amount:   "123456",
					},
				),
			},
			"", "",
			true,
		},
		{
			"invalid recipient",
			// nolint: exhaustruct
			channeltypes.Packet{
				Data: transfertypes.ModuleCdc.MustMarshalJSON(
					// nolint: exhaustruct
					&transfertypes.FungibleTokenPacketData{
						Sender:   "cosmos1qql8ag4cluz6r4dz28p3w00dnc9w8ueulg2gmc",
						Receiver: "althea1",
						Amount:   "123456",
					},
				),
			},
			"", "",
			true,
		},
		{
			"valid - cosmos sender, althea recipient",
			// nolint: exhaustruct
			channeltypes.Packet{
				Data: transfertypes.ModuleCdc.MustMarshalJSON(
					// nolint: exhaustruct
					&transfertypes.FungibleTokenPacketData{
						Sender:   "cosmos1qql8ag4cluz6r4dz28p3w00dnc9w8ueulg2gmc",
						Receiver: "althea1x2w87cvt5mqjncav4lxy8yfreynn273x9h93zp",
						Amount:   "123456",
					},
				),
			},
			"althea1qql8ag4cluz6r4dz28p3w00dnc9w8ueut20xha",
			"althea1x2w87cvt5mqjncav4lxy8yfreynn273x9h93zp",
			false,
		},
		{
			"valid - althea sender, cosmos recipient",
			// nolint: exhaustruct
			channeltypes.Packet{
				Data: transfertypes.ModuleCdc.MustMarshalJSON(
					// nolint: exhaustruct
					&transfertypes.FungibleTokenPacketData{
						Sender:   "althea1x2w87cvt5mqjncav4lxy8yfreynn273x9h93zp",
						Receiver: "cosmos1qql8ag4cluz6r4dz28p3w00dnc9w8ueulg2gmc",
						Amount:   "123456",
					},
				),
			},
			"althea1x2w87cvt5mqjncav4lxy8yfreynn273x9h93zp",
			"althea1qql8ag4cluz6r4dz28p3w00dnc9w8ueut20xha",
			false,
		},
		{
			"valid - osmosis sender, althea recipient",
			// nolint: exhaustruct
			channeltypes.Packet{
				Data: transfertypes.ModuleCdc.MustMarshalJSON(
					// nolint: exhaustruct
					&transfertypes.FungibleTokenPacketData{
						Sender:   "osmo1qql8ag4cluz6r4dz28p3w00dnc9w8ueuhnecd2",
						Receiver: "althea1x2w87cvt5mqjncav4lxy8yfreynn273x9h93zp",
						Amount:   "123456",
					},
				),
			},
			"althea1qql8ag4cluz6r4dz28p3w00dnc9w8ueut20xha",
			"althea1x2w87cvt5mqjncav4lxy8yfreynn273x9h93zp",
			false,
		},
	}

	for _, tc := range testCases {
		sender, recipient, _, _, err := GetTransferSenderRecipient(tc.packet)

		if tc.expError {
			require.Error(t, err, tc.name)
		} else {
			require.NoError(t, err, tc.name)
			require.Equal(t, tc.expSender, sender.String())
			require.Equal(t, tc.expRecipient, recipient.String())
		}
	}
}

func TestGetTransferAmount(t *testing.T) {
	testCases := []struct {
		name      string
		packet    channeltypes.Packet
		expAmount string
		expError  bool
	}{
		{
			"empty packet",
			// nolint: exhaustruct
			channeltypes.Packet{},
			"",
			true,
		},
		{
			"invalid packet data",
			// nolint: exhaustruct
			channeltypes.Packet{
				Data: ibctesting.MockFailPacketData,
			},
			"",
			true,
		},
		{
			"invalid amount - empty",
			// nolint: exhaustruct
			channeltypes.Packet{
				Data: transfertypes.ModuleCdc.MustMarshalJSON(
					// nolint: exhaustruct
					&transfertypes.FungibleTokenPacketData{
						Sender:   "cosmos1qql8ag4cluz6r4dz28p3w00dnc9w8ueulg2gmc",
						Receiver: "althea1x2w87cvt5mqjncav4lxy8yfreynn273x9h93zp",
						Amount:   "",
					},
				),
			},
			"",
			true,
		},
		{
			"invalid amount - non-int",
			// nolint: exhaustruct
			channeltypes.Packet{
				Data: transfertypes.ModuleCdc.MustMarshalJSON(
					// nolint: exhaustruct
					&transfertypes.FungibleTokenPacketData{
						Sender:   "cosmos1qql8ag4cluz6r4dz28p3w00dnc9w8ueulg2gmc",
						Receiver: "althea1x2w87cvt5mqjncav4lxy8yfreynn273x9h93zp",
						Amount:   "test",
					},
				),
			},
			"test",
			true,
		},
		{
			"valid",
			// nolint: exhaustruct
			channeltypes.Packet{
				Data: transfertypes.ModuleCdc.MustMarshalJSON(
					// nolint: exhaustruct
					&transfertypes.FungibleTokenPacketData{
						Sender:   "cosmos1qql8ag4cluz6r4dz28p3w00dnc9w8ueulg2gmc",
						Receiver: "althea1x2w87cvt5mqjncav4lxy8yfreynn273x9h93zp",
						Amount:   "10000",
					},
				),
			},
			"10000",
			false,
		},
	}

	for _, tc := range testCases {
		amt, err := GetTransferAmount(tc.packet)
		if tc.expError {
			require.Error(t, err, tc.name)
		} else {
			require.NoError(t, err, tc.name)
			require.Equal(t, tc.expAmount, amt)
		}
	}
}
